+++ packages/plugin-autocapture-browser/src/autocapture/track-dead-click.ts
@@ -40,7 +40,6 @@ export function trackDeadClick({
   const changeObservables: Array<
     AllWindowObservables[ObservablesEnum.MutationObservable] | AllWindowObservables[ObservablesEnum.NavigateObservable]
   > = [mutationObservable];
-  /* istanbul ignore next */
   if (navigateObservable) {
     changeObservables.push(navigateObservable);
   }
@@ -49,7 +48,7 @@ export function trackDeadClick({
   const actionClicks = filteredClickObservable.pipe(
     mergeMap((click) => {
       // Create a timer that emits after 500ms
-      let timeoutId: NodeJS.Timeout;
+      let timeoutId: ReturnType<typeof setTimeout>;
       const timer = new Observable<typeof click>((subscriber: Subscriber<typeof click>) => {
         timeoutId = setTimeout(() => subscriber.next(click), DEAD_CLICK_TIMEOUT);
 
@@ -77,8 +76,7 @@ export function trackDeadClick({
       X: (actionClick.event as MouseEvent).clientX,
       Y: (actionClick.event as MouseEvent).clientY,
     };
-    /* istanbul ignore next */
-    amplitude?.track(
+    amplitude.track(
       AMPLITUDE_ELEMENT_DEAD_CLICKED_EVENT,
       {
         ...getEventProperties('click', actionClick.closestTrackedAncestor),