+++ packages/plugin-autocapture-browser/src/frustration-plugin.ts
@@ -34,18 +34,14 @@ export const frustrationPlugin = (
     },
   },
 ): BrowserEnrichmentPlugin => {
-  console.log('!!!options', options);
   const name = constants.FRUSTRATION_PLUGIN_NAME;
   const type = 'enrichment';
-
-  // TODO: add visualTagging stuff here too
 
   const subscriptions: Subscription[] = [];
 
   const rageCssSelectors = options.rageClicks?.cssSelectorAllowlist ?? DEFAULT_RAGE_CLICK_ALLOWLIST;
   const deadCssSelectors = options.deadClicks?.cssSelectorAllowlist ?? DEFAULT_DEAD_CLICK_ALLOWLIST;
 
-  /* istanbul ignore next */
   const dataAttributePrefix = options.dataAttributePrefix ?? DEFAULT_DATA_ATTRIBUTE_PREFIX;
 
   // combine the two selector lists to determine which clicked elements should be filtered
@@ -59,9 +55,7 @@ export const frustrationPlugin = (
         return addAdditionalEventProperties(
           click,
           'click',
-          /* istanbul ignore next */
           combinedCssSelectors,
-          /* istanbul ignore next */
           dataAttributePrefix,
           true, // capture when cursor is pointer
         );
@@ -83,15 +77,7 @@ export const frustrationPlugin = (
 
     // Track DOM Mutations
     const enrichedMutationObservable = createMutationObservable().pipe(
-      map((mutation) =>
-        addAdditionalEventProperties(
-          mutation,
-          'mutation',
-          /* istanbul ignore next */
-          combinedCssSelectors,
-          dataAttributePrefix,
-        ),
-      ),
+      map((mutation) => addAdditionalEventProperties(mutation, 'mutation', combinedCssSelectors, dataAttributePrefix)),
       share(),
     );
 
@@ -136,12 +122,10 @@ export const frustrationPlugin = (
     config?.loggerProvider?.log(`${name} has been successfully added.`);
   };
 
-  /* istanbul ignore next */
   const execute: BrowserEnrichmentPlugin['execute'] = async (event) => {
     return event;
   };
 
-  /* istanbul ignore next */
   const teardown = async () => {
     for (const subscription of subscriptions) {
       subscription.unsubscribe();