+++ packages/segment-session-replay-plugin/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.0.0-beta.12](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.11...@amplitude/segment-session-replay-plugin@0.0.0-beta.12) (2025-06-25)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
 # [0.0.0-beta.11](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.10...@amplitude/segment-session-replay-plugin@0.0.0-beta.11) (2025-06-17)
 
 **Note:** Version bump only for package @amplitude/segment-session-replay-plugin